+++ src/View/Helper/AbstractMenuHelper.php
@@ -11,12 +11,23 @@ declare(strict_types=1);
  * @copyright   Copyright (c) Mirko Pagliai
  * @link        https://github.com/mirko-pagliai/me-cms
  * @license     https://opensource.org/licenses/mit-license.php MIT License
- * @var string $email
- * @var string $firstName
- * @var string $lastName
- * @var string $message
+ * @since       2.31.9
  */
 
-echo __d('me_cms', 'Email from {0}', sprintf('%s %s (%s)', $firstName, $lastName, $email));
+namespace MeCms\View\Helper;
 
-echo $message;
+use Cake\View\Helper;
+
+/**
+ * `AbstractMenuHelper` for all `MenuHelper` classes
+ * @see \MeCms\View\Helper\MenuBuilderHelper::generate() for more information
+ * @property \MeCms\View\Helper\IdentityHelper $Identity
+ */
+abstract class AbstractMenuHelper extends Helper
+{
+    /**
+     * Helpers
+     * @var array
+     */
+    public $helpers = ['MeCms.Identity'];
+}