+++ tests/test_app/TestApp/Plugin/TestPluginThree/src/View/Helper/MenuHelper.php
@@ -1,0 +1,10 @@
+<?php
+declare(strict_types=1);
+
+namespace TestPluginThree\View\Helper;
+
+use Cake\View\Helper;
+
+class MenuHelper extends Helper
+{
+}